+++ src/ThemeChooser.js
@@ -1,60 +1,92 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from "react";
+import PropTypes from "prop-types";
 
-function capitalize(s) {
-  return s.charAt(0).toUpperCase() + s.substring(1);
-}
+const capitalize = (s) => s.charAt(0).toUpperCase() + s.substring(1)
 
 class ThemeChooser extends React.Component {
   constructor(props, context) {
-    super(props);
+    super(props, context);
     this.onSelect = this.onSelect.bind(this);
-    this.doCallback = this.doCallback.bind(this);
 
     // get themes from context and sort them for display
     this.themes = [];
+
     context.themes.forEach(theme => {
       this.themes.push(theme);
     });
-    this.themes.sort();
-  }
 
-  doCallback() {
+    this.themes.sort();
+
+    this.state = {
+      currentTheme: this.context.currentTheme || "",
+      defaultTheme: this.context.defaultTheme
+    }
   }
 
   onSelect(e) {
     e.preventDefault();
-    let chosenTheme = e.target.getAttribute('data-theme');
-    this.context.themeSwitcher.load(chosenTheme);
+    this.setState({
+      currentTheme: e.target.getAttribute("data-theme")
+    }, () => {
+      this.context.themeSwitcher.load(this.state.currentTheme);
+      this.props.change(this.state.currentTheme)
+    })
   }
 
   render() {
-    let style = this.props.style || {};
-    const menu =
-      <div className="dropdown dropdown-menu-right" style={style}>
-        <button className="btn btn-default dropdown-toggle" type="button" id="theme-menu"
-                  data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
-          Select Theme
-          &nbsp;<span className="caret"></span>
-      </button>
+    let themes = this.themes.map(theme => {
+      return (
+        <li key={ theme }>
+          <a
+            href="#"
+            className={ "dropdown-item" + (theme === this.state.currentTheme ? " active" : "") }
+            data-theme={ theme }
+            onClick={ this.onSelect }
+          >
+            { theme === this.state.defaultTheme ? "* " : "" }{ capitalize(theme) }
+          </a>
+        </li>
+      )
+    })
+
+    return (
+      <div className="dropdown dropdown-menu-right" style={ this.props.style }>
+        <button
+          id="theme-menu"
+          className={ "btn btn-default dropdown-toggle" + (this.props.size === "" ? "" : " btn-" + this.props.size) }
+          type="button"
+          data-toggle="dropdown"
+          aria-haspopup="true"
+          aria-expanded="true"
+        >
+          Theme
+        </button>
+
         <ul className="dropdown-menu">
-          {this.themes.map((theme) => {
-            var active = (theme === this.context.currentTheme ? 'active' : '');
-            return <li key={theme} className={active}>
-              <a href="#" data-theme={theme} onClick={this.onSelect}>{capitalize(theme)}</a>
-            </li>
-          })}
+          { themes }
         </ul>
       </div>
-
-    return menu;
+    )
   }
 }
 
 ThemeChooser.contextTypes = {
+  defaultTheme: PropTypes.string,
   themeSwitcher: PropTypes.object,
   themes: PropTypes.array,
   currentTheme: PropTypes.string
 };
 
-export { ThemeChooser };
+ThemeChooser.propTypes = {
+  style: PropTypes.object,
+  size: PropTypes.oneOf(["xs", "sm", "md", "lg"]),
+  change: PropTypes.func
+};
+
+ThemeChooser.defaultProps = {
+  style: {},
+  size: "",
+  change: (t) => {}
+};
+
+export default ThemeChooser;