+++ src/themes/index.js
@@ -1,0 +1,26 @@
+let validThemes = [
+	'cerulean',
+	 'cosmo',
+	 'cyborg',
+	 'darkly',
+	 'flatly',
+	 'journal',
+	 'litera',
+	 'lumen',
+	 'lux',
+	 'materia',
+	 'minty',
+	 'pulse',
+	 'sandstone',
+	 'simplex',
+	 'sketchy',
+	 'slate',
+	 'solar',
+	 'spacelab',
+	 'superhero',
+	 'united',
+	 'yeti'
+]
+export {
+	validThemes
+}