+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [1.0.7]
+* Themes are saved/loaded from js files for faster switching
+* Theme specific files (fonts) are saved locally for faster switching 
+
 ### [1.0.6]
 * ThemeChooser dropdown current theme shown as active.
 * Remove console.log